# container/lru: Implement type safe generic LRUs.