--- a/container/lru/go.mod
+++ b/container/lru/go.mod
@@ -1,0 +1,3 @@
+module github.com/decred/dcrd/container/lru
+
+go 1.18