--- a/go.mod
+++ b/go.mod
@@ -15,6 +15,7 @@ require (
 	github.com/decred/dcrd/chaincfg/v3 v3.2.1
 	github.com/decred/dcrd/connmgr/v3 v3.1.2
 	github.com/decred/dcrd/container/apbf v1.0.1
+	github.com/decred/dcrd/container/lru v1.0.0
 	github.com/decred/dcrd/crypto/blake256 v1.0.1
 	github.com/decred/dcrd/crypto/ripemd160 v1.0.2
 	github.com/decred/dcrd/database/v3 v3.0.2
@@ -69,6 +70,7 @@ replace (
 	github.com/decred/dcrd/chaincfg/v3 => ./chaincfg
 	github.com/decred/dcrd/connmgr/v3 => ./connmgr
 	github.com/decred/dcrd/container/apbf => ./container/apbf
+	github.com/decred/dcrd/container/lru => ./container/lru
 	github.com/decred/dcrd/crypto/blake256 => ./crypto/blake256
 	github.com/decred/dcrd/crypto/ripemd160 => ./crypto/ripemd160
 	github.com/decred/dcrd/database/v3 => ./database